requirements/base.txt
argon2-cffi==23.1.0
celery==5.4.0
channels-redis==4.1.0
channels==4.0.0
daphne==4.0.0
django-celery-beat==2.5.0
django-environ==0.11.2
django-redis==5.3.0
django==4.2.9
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
flower==2.0.1
phonenumbers==8.13.40
Pillow==10.1.0
redis==5.0.1
twilio==9.2.3
twisted[tls,http2]==24.3.0
whitenoise==6.7.0
uvicorn[standard]==0.30.3
uvicorn-worker==0.2.0
